requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
datasets>=2.14.0
Pillow>=10.0.0

# Visualization & metrics
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Progress & logging
tqdm>=4.65.0
pyyaml>=6.0

# Additional utilities
opencv-python>=4.8.0
scipy>=1.11.0
